uv/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
filetype==1.2.0
flask==3.1.0
fonttools==4.57.0
geopandas==1.0.1
google-ai-generativelanguage==0.6.17
google-api-core[grpc]==2.24.2
google-auth==2.39.0
google-custom-search==3.0.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.23
langchain-core==0.3.51
langchain-google-genai==2.0.11
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langgraph==0.3.25
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.24
markupsafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
openai==1.70.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.2
pydantic-core==2.33.1
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
werkzeug==3.1.3
xxhash==3.5.0
zstandard==0.23.0